refactor : moved get database  to utils
