+++ internals/handlers/handlers.go
@@ -12,6 +12,7 @@ import (
 	"time"
 
 	"Heal/internals/renders"
+	"Heal/utils"
 
 	"github.com/google/uuid"
 	"github.com/joho/godotenv"
@@ -406,38 +407,6 @@ func LoginHandler(w http.ResponseWriter, r *http.Request) {
 	})
 }
 
-func Getdb(name string) (*sql.DB, error) {
-	db, err := sql.Open("sqlite3", name)
-	if err != nil {
-		log.Fatalf("Failed to connect to database: %v", err)
-	}
-	defer db.Close()
-
-	// Create the users table if it doesn't exist
-	createTableQuery := `
-	CREATE TABLE IF NOT EXISTS users (
-		id INTEGER PRIMARY KEY AUTOINCREMENT,
-		username TEXT UNIQUE NOT NULL,
-		password TEXT NOT NULL
-	);`
-	if _, err := db.Exec(createTableQuery); err != nil {
-		log.Fatalf("Failed to create table: %v", err)
-	}
-
-	createsessionquerry := `
-	CREATE TABLE IF NOT EXISTS sessions (
-		session_id TEXT PRIMARY KEY,
-		user_id INTEGER NOT NULL,
-		expires_at DATETIME NOT NULL,
-		FOREIGN KEY (user_id) REFERENCES users(id)
-	);`
-
-	if _, err := db.Exec(createsessionquerry); err != nil {
-		log.Fatalf("Failed to create table: %v", err)
-	}
-
-	return db, nil
-}
 
 // singup handler
 func SignupHandler(w http.ResponseWriter, r *http.Request) {
@@ -445,7 +414,7 @@ func SignupHandler(w http.ResponseWriter, r *http.Request) {
 	if err != nil {
 		log.Fatalf("Failed to connect to database: %v", err)
 	}
-	Getdb("./Heal.db")
+	utils.Getdb("./Heal.db")
 	defer db.Close()
 	if r.Method != http.MethodPost {
 		http.Error(w, "Invalid request method", http.StatusMethodNotAllowed)
